+++ WMFData/Tests/WMFDataTests/WMFWatchlistDataControllerTests.swift
@@ -369,54 +369,6 @@ final class WMFWatchlistDataControllerTests: XCTestCase {
          }
      }
     
-    func testFetchWatchStatus() {
-         let controller = WMFWatchlistDataController()
-
-         let expectation = XCTestExpectation(description: "Fetch Watch Status")
-         var statusToTest: WMFPageWatchStatus?
-        controller.fetchWatchStatus(title: "Cat", project: enProject) { result in
-             switch result {
-             case .success(let status):
-                 statusToTest = status
-             case .failure(let error):
-                 XCTFail("Failure fetching watch status: \(error)")
-             }
-             expectation.fulfill()
-         }
-
-         guard let statusToTest else {
-             XCTFail("Missing statusToTest")
-             return
-         }
-
-         XCTAssertTrue(statusToTest.watched)
-         XCTAssertNil(statusToTest.userHasRollbackRights)
-     }
-
-     func testFetchWatchStatusWithRollbackRights() {
-         let controller = WMFWatchlistDataController()
-
-         let expectation = XCTestExpectation(description: "Fetch Watch Status")
-         var statusToTest: WMFPageWatchStatus?
-         controller.fetchWatchStatus(title: "Cat", project: enProject, needsRollbackRights: true) { result in
-             switch result {
-             case .success(let status):
-                 statusToTest = status
-             case .failure(let error):
-                 XCTFail("Failure fetching watch status: \(error)")
-             }
-             expectation.fulfill()
-         }
-
-         guard let statusToTest else {
-             XCTFail("Missing statusToTest")
-             return
-         }
-
-         XCTAssertFalse(statusToTest.watched)
-         XCTAssertTrue((statusToTest.userHasRollbackRights ?? false))
-     }
-    
     func testPostRollbackArticle() {
         let controller = WMFWatchlistDataController()
